backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy[asyncio]==2.0.30
asyncio==3.4.3
aiosqlite==0.20.0

# Validation
pydantic[email]==2.7.1
pydantic-settings==2.2.1

# Config
python-dotenv==1.0.1

# Google AI & Services
google-generativeai==0.5.4
opencage==2.4.0

# HTTP client
httpx==0.27.0

# Utils
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9

# Logging
structlog==24.1.0
